crawler: fix flags, add db.go, remove discv4.go and discv5.go
